Move base functionality for foreach_scaling benchmark to header file
